Refactor move to hooks
